+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 [Full Changelog](https://github.com/ppfeufer/eve-online-fitting-manager/compare/v0.13.1...HEAD)
 ### Changed
 - Ship types overview now as well in the new look
+- Moved some code into its own template
 
 ## [v0.13.1](https://github.com/ppfeufer/eve-online-fitting-manager/releases/tag/v0.13.1) - 2017-10-11
 [Full Changelog](https://github.com/ppfeufer/eve-online-fitting-manager/compare/v0.13...v0.13.1)